Split up the SGX shim's handler
This makes it easier to reason about.
